Improve protection against invalid device selection

+++ prepare-usb.nu
@@ -2,10 +2,15 @@
 
 def main [tag: string] {
   lsblk
-  let dev = $"/dev/(input 'Enter device name (eg. sdb): ')"
-  ls $dev
-  do --ignore-program-errors { sudo umount --quiet $'($dev)*' }
-  curl --location --fail $'https://github.com/nicholaslawton/nickos/releases/download/($tag)/nickos.iso' --output nickos.iso
-  sudo dd if=nickos.iso $'of=($dev)' bs=4M conv=fsync status=progress
-  rm nickos.iso
+  let name = input 'Enter device name (eg. sdb): '
+  let dev = (ls --directory $'/dev/($name)' | first)
+  if $dev.name == $'/dev/($name)' and $dev.type == 'block device' {
+    do --ignore-program-errors { sudo umount --quiet $'($dev)*' }
+    curl --location --fail $'https://github.com/nicholaslawton/nickos/releases/download/($tag)/nickos.iso' --output nickos.iso
+    sudo dd if=nickos.iso $'of=($dev.name)' bs=4M conv=fsync status=progress
+    rm nickos.iso
+  } else {
+    echo $'invalid device - "($name)" does not identify a block device'
+    echo $dev
+  }
 }
